+++ docs/index.js
@@ -2,9 +2,8 @@ import '@babel/polyfill'
 import React from 'react'
 import {render} from 'react-dom'
 import {HashRouter, Switch} from 'react-router-dom'
-import {ThemeProvider as UIThemeProvider} from 'rambler-ui/theme'
 import pageList from 'docs/page-list'
-import {ThemeProvider as DocsThemeProvider} from 'docs/utils/theming'
+import {ThemeProvider} from 'docs/utils/theming'
 import App from 'docs/components/App'
 import Route from 'docs/components/Route'
 import SideNav from 'docs/components/SideNav'
@@ -25,23 +24,21 @@ const flattenRoutes = (routes, acc = []) =>
   )
 
 const root = (
-  <UIThemeProvider>
-    <DocsThemeProvider>
-      <HashRouter>
-        <App>
-          <SideNav index={pageList} />
-          <div>
-            <Switch>
-              {flattenRoutes(pageList)}
-              <Route key="/" path="/" />
-            </Switch>
-            <Informer />
-            <Footer />
-          </div>
-        </App>
-      </HashRouter>
-    </DocsThemeProvider>
-  </UIThemeProvider>
+  <ThemeProvider>
+    <HashRouter>
+      <App>
+        <SideNav index={pageList} />
+        <div>
+          <Switch>
+            {flattenRoutes(pageList)}
+            <Route key="/" path="/" />
+          </Switch>
+          <Informer />
+          <Footer />
+        </div>
+      </App>
+    </HashRouter>
+  </ThemeProvider>
 )
 
 render(root, document.getElementById('root'))