+++ src/theme/index.js
@@ -10,7 +10,6 @@ export {
 
 export {default as ThemeProvider} from './ThemeProvider'
 export {default as withStyles} from './with-styles'
-export {default as createUseStyles} from './create-use-styles'
 
 export {default as colors} from './colors'
 export {default as i18n} from './i18n'